// backend/src/visits/dto/create-visit.dto.d.ts
export declare class CreateVisitDto {
    visitDate?: string;
    visitType: string;
    notes?: string;
    price?: number;
    nextReminderDate?: string;
    isReminderEnabled?: boolean;
    temperature?: number;
    weight?: number;
    weightUnit?: string;
    heartRate?: number;
    respiratoryRate?: number;
}